requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.3
backcall==0.1.0
beautifulsoup4==4.6.0
Brlapi==0.6.6
certifi==2018.11.29
cffi==1.12.2
chardet==3.0.4
command-not-found==0.3
cryptography==2.6.1
cssselect==1.0.3
cupshelpers==1.0
decorator==4.4.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==2.1.7
django-debug-toolbar==1.11
django-extensions==2.1.6
django-guardian==1.5.0
django-ipware==2.1.0
django-mailer==1.2.5
django-modelcluster==4.3
django-polymorphic==2.0.3
django-render-block==0.5
django-simple-history==2.7.0
django-taggit==0.23.0
django-treebeard==4.3
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
draftjs-exporter==2.1.5
filelock==3.0.10
future==0.17.1
google-api-python-client==1.5.0
howdoi==1.1.14
html5lib==1.0.1
httpie==0.9.8
httplib2==0.9.2
idna==2.8
ipython==7.4.0
ipython-genutils==0.2.0
isort==4.3.15
jedi==0.13.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
lockfile==0.12.2
louis==3.5.0
lxml==4.3.2
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==6.0.0
netifaces==0.10.4
numpy==1.13.3
oauth==1.0.1
oauth2client==2.2.0
olefile==0.45.1
opencv-python==4.0.0.21
pandas==0.24.2
parso==0.3.4
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.9.0
prompt-toolkit==2.0.9
protobuf==3.0.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycairo==1.16.2
pycodestyle==2.5.0
pycountry==18.12.8
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
pygobject==3.26.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pymacaroons==0.13.0
PyNaCl==1.1.2
pyquery==1.4.0
pyRFC3339==1.0
pytest==4.3.1
python-apt==1.6.3+ubuntu1
python-dateutil==2.8.0
python-debian==0.1.32
pytracking==0.2.0
pytz==2018.9
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.21.0
requests-cache==0.5.0
requests-unixsocket==0.1.5
rest-pandas==1.1.0
rsa==4.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.12.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
toml==0.10.0
tox==3.7.0
traitlets==4.3.2
typed-ast==1.3.1
ua-parser==0.8.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.0.23
uritemplate==0.6
urllib3==1.24.1
usb-creator==0.3.3
user-agents==1.1.0
virtualenv==16.4.3
wadllib==1.3.2
wagalytics==0.8.1
wagtail==2.4
wagtail-metadata==2.0.1
wagtail-personalisation==0.12.1
wagtailfontawesome==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
Willow==1.1
wrapt==1.11.1
xkit==0.0.0
zope.interface==4.3.2
